design_patterns/002_observer/solutions/python/ImageProcessing/requirements.txt
reactivex
colorama
pillow
